requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0
python-dotenv>=0.19.0
pydantic>=1.8.0
python-multipart>=0.0.5
aiofiles>=0.8.0
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=0.24.0
joblib>=1.0.0
plotly>=5.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
python-jose[cryptography]
passlib[bcrypt]
reportlab==4.1.0
